==> notebooks/neural_response_generator/00_conversational.ipynb <==
import sys
root_path = "/Users/sa/CS/stanford/cs229/Project/Neural-Response-Generator-augmented-with-psychological-personalities/"

import pandas as pd
import autoreload
import sys
sys.path.append(root_path + "library/")

%load_ext autoreload
%autoreload 2

# essentials
import numpy as np
from tqdm.notebook import tqdm
import pickle
import json

# Formating Movie data

import csv

with open('/Users/sa/Downloads/cornell/movie_conversations.txt', 'r') as f:
    content = [row.split(" +++$+++ ") for row in f.readlines()]
    
movie_conversation = pd.DataFrame(data=content, columns=["char1Id", "char2Id", "movieId", "lines"])

movie_conversation.to_csv("../../data/processed/final/movie_conversation.csv")

import csv

with open('/Users/sa/Downloads/cornell/movie_lines.txt', 'r', errors='replace') as f:
    #f.readlines()
    content = [row.split(" +++$+++ ") for row in f.readlines()]
    
movie_lines = pd.DataFrame(data=content, columns=["lineId", "charId", "movieId", "charName", "line"])

movie_lines.to_csv("../../data/processed/final/movie_lines.csv")

import csv

with open('/Users/sa/Downloads/cornell/movie_titles_metadata.txt', 'r', errors='replace') as f:
    #f.readlines()
    content = [row.split(" +++$+++ ") for row in f.readlines()]
    
movie_titles = pd.DataFrame(data=content, columns=["movieId", "movieTitle", "movieYr", "IMDBr", "IMDBv", "genres"])

movie_titles.to_csv("../../data/processed/final/movie_titles.csv")

movie_titles.sort_values(by='IMDBr',ascending=False)[:30]

movie_script = movie_conversation.loc[movie_conversation['movieId'].isin(list(movie_titles.sort_values(by='IMDBr',ascending=False).movieId[:50]))].copy()

import ast

#movie_script['lines']
contexted = []
for itr, rows in tqdm(movie_script.iterrows(), total=movie_script.shape[0], mininterval=20):
    line_row = []
    for lineId in ast.literal_eval(rows.lines):
        line_row.append(movie_lines.loc[movie_lines['lineId']==lineId].line)
    tmp = []
    for i in line_row:
        for j in i:
            tmp.append(j[:-1])
    movie_script.at[itr, 'lines'] = tmp
    contexted.append(tmp)
    

movie_script.to_csv('../../data/intermediate/movie_script_elong.csv')

movie_script
