==> notebooks/neural_response_generator/01_conversational.ipynb <==
import sys
root_path = "/Users/sa/CS/stanford/cs229/Project/Neural-Response-Generator-augmented-with-psychological-personalities/"

import pandas as pd
import autoreload
import sys
sys.path.append(root_path + "library/")

%load_ext autoreload
%autoreload 2

# essentials
import numpy as np
from tqdm.notebook import tqdm
import pickle
import json

script = pd.read_csv(root_path + '/data/intermediate/movie_script_elong.csv')

script.head()

import ast

#contexted = []
groupid = 0
script_df = pd.DataFrame(columns=["groupId", "line"])
for itr, row in tqdm(script.iterrows(), total=script.shape[0], mininterval=20):
    for line in ast.literal_eval(row.lines):
        script_df.loc[script_df.shape[0]+1] = [groupid, line]
    groupid += 1
    #contexted.append(ast.literal_eval(row.lines))

script_df = script_df.reset_index(drop=True).copy()

script_df

# Prepare the data in conversational style with context

# load personality classifier model

from joblib import load

name = 'mlpc'
clf = load(root_path + 'models/final/' + name +'.joblib') 

import pickle
cvectorizer = pickle.load(open(root_path + "models/final/cvectorizer.pkl", "rb"))
vectorizer = pickle.load(open(root_path + "models/final/vectorizer.pkl", "rb"))
svd = pickle.load(open(root_path + "models/final/tsvd.pkl", "rb"))
label_dict = pickle.load(open(root_path + "models/final/label_dict.pkl", "rb"))

contexted = []

n = 7

for i in range(n, len(script_df['line'])):
    row = []
    prev = i - 1 - n
    c_grp = script_df['groupId'][i]
    for j in range(i, prev, -1):
        if script_df['groupId'][j] != c_grp:
            break
        row.append(script_df['line'][j])
    if len(row) == n+1:
        contexted.append(row) 

columns = ['response', 'context'] 
columns = columns + ['context/'+str(i) for i in range(n-1)]

df = pd.DataFrame.from_records(contexted, columns=columns)
df.head(5)

personality = []

X_vec = vectorizer.transform(cvectorizer.transform(list(df.response)))
X_vec_svd = svd.transform(X_vec)
personality = np.asarray([label_dict[i] for i in clf.predict(X_vec_svd)])

from sklearn.model_selection import train_test_split

trn_df, val_df = train_test_split(df.loc[personality == 'ISFJ'], test_size = 0.1)
#trn_df, val_df = train_test_split(df, test_size = 0.1) # fullset training
print (trn_df.shape)
trn_df.head()

trn_df.to_csv(root_path + "data/processed/final/trn_df.csv")
val_df.to_csv(root_path + "data/processed/final/val_df.csv")
